+++ finalwork/Core/App/step.h
@@ -3,5 +3,5 @@
 
 void step_init();
 void step_dir_set(uint8_t dir);
-void step_set(uint32_t round);
+void step_set(float round);
 void delay_us(uint16_t us);
